Search View Commit

+++ app/src/main/java/com/example/cryptotrack/data/source/dto/CodeAdditionsDeletions4Weeks.kt
@@ -1,6 +1,6 @@
 package com.example.cryptotrack.data.source.dto
 
 data class CodeAdditionsDeletions4Weeks(
-    val additions: Int,
-    val deletions: Int
+    val additions: Double,
+    val deletions: Double
 )
