+++ app/src/main/java/com/example/cryptotrack/data/source/dto/ConvertedLast.kt
@@ -3,5 +3,5 @@ package com.example.cryptotrack.data.source.dto
 data class ConvertedLast(
     val btc: Double,
     val eth: Double,
-    val usd: Int
+    val usd: Double
 )
