+++ app/src/main/java/com/example/cryptotrack/data/source/dto/PublicInterestStats.kt
@@ -1,6 +1,6 @@
 package com.example.cryptotrack.data.source.dto
 
 data class PublicInterestStats(
-    val alexa_rank: Int,
+    val alexa_rank: Double,
     val bing_matches: Any
 )
